client-app/requirements.txt
ultralytics
requests
google-generativeai==0.3.1
opencv-python-headless
torch
Pillow
numpy